Rename 'StateArchiveMode' to 'ArchiveMode'
