+++ packages/beacon-node/test/perf/chain/produceBlock/produceBlockBody.test.ts
@@ -10,7 +10,7 @@ import {BeaconChain} from "../../../../src/chain/index.js";
 import {BlockType, produceBlockBody} from "../../../../src/chain/produceBlock/produceBlockBody.js";
 import {Eth1ForBlockProductionDisabled} from "../../../../src/eth1/index.js";
 import {ExecutionEngineDisabled} from "../../../../src/execution/engine/index.js";
-import {BeaconDb, StateArchiveMode} from "../../../../src/index.js";
+import {BeaconDb, ArchiveMode} from "../../../../src/index.js";
 import {testLogger} from "../../../utils/logger.js";
 
 const logger = testLogger();
@@ -36,7 +36,7 @@ describe("produceBlockBody", () => {
         skipCreateStateCacheIfAvailable: true,
         archiveStateEpochFrequency: 1024,
         minSameMessageSignatureSetsToBatch: 32,
-        stateArchiveMode: StateArchiveMode.Frequency,
+        archiveMode: ArchiveMode.Frequency,
       },
       {
         config: state.config,
