+++ packages/beacon-node/test/perf/chain/verifyImportBlocks.test.ts
@@ -11,7 +11,7 @@ import {AttestationImportOpt, BlockSource, getBlockInput} from "../../../src/cha
 import {BeaconChain} from "../../../src/chain/index.js";
 import {Eth1ForBlockProductionDisabled} from "../../../src/eth1/index.js";
 import {ExecutionEngineDisabled} from "../../../src/execution/engine/index.js";
-import {BeaconDb, StateArchiveMode} from "../../../src/index.js";
+import {BeaconDb, ArchiveMode} from "../../../src/index.js";
 import {linspace} from "../../../src/util/numpy.js";
 import {testLogger} from "../../utils/logger.js";
 
@@ -85,7 +85,7 @@ describe.skip("verify+import blocks - range sync perf test", () => {
           skipCreateStateCacheIfAvailable: true,
           archiveStateEpochFrequency: 1024,
           minSameMessageSignatureSetsToBatch: 32,
-          stateArchiveMode: StateArchiveMode.Frequency,
+          archiveMode: ArchiveMode.Frequency,
         },
         {
           config: state.config,
